Replace DatadogSink implementation with alternative
In the previous version, everything was added to a giant (persistent) StringBuilder, we then call ToString(), and convert the string to a byte array (which is then thrown away). The string and utf8 array are ultimately garbage.

In this version, we keep a smaller persistent StringBuilder (one for each log, instead of all logs), convert that to a string, and write the bytes to a persistent byte array. Only the intermediate string is garbage.

The benefit should be that we have less chance of the string/byte array ending up on the LOH, and we can reuse the UTF8 array