Handle errors when pushing telemetry
There are several cases where we _expect_ telemetry to fail to send. The most common will likely when the agent version isn't new enough, so we will get a 404 when sending.

To handle these "fatal" errors in addition to normal transient errors, we use the following approach.

- If we receive two 404s in a row from the endpoint, _without_ ever getting a success, we stop trying to push telemetry.
- If we receive five errors in a row from the endpoint, at any time, we stop trying to push telemetry.

If we have a transient error, we don't retry immediately. Instead, we hold onto the config/dependencies/integrations data, and use that to build the next request, unless these change in the mean-time. Currently, this is fixed at 1 minute between requests, as per the spec.

Note that for simplicity, we're currently still _recording_ telemetry, we're just not doing anything with it once we have too many failures. A subsequent improvement would be to trigger a replacement of the TelemetryController with the NullTelemetryController, but that's a little more complex to do.