Update instrumentation definitions generator to also implement a `GetIntegrationId()` method
This can be used at runtime by calltarget instrumentation to determine which `IntegrationId` an executing method is associated with. There may be other ways of doing it as well, but given this only needs to be done once per instrumentation type, I think this is sufficient.

Note that the pattern matching code produces lots of duplicates, but the compiler is smart enough to notice they're all the same and merge them into one.